pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_qus_analyzer"
version = "0.1.0"
description = "A quality user story analyzer using LLM"
authors = [
    {name = "Waffiq Maaroja", email = "waffiqmaaroja1998@gmail.com"},
    {name = "Wicaksono Leksono", email = "wcksnlxn@gmail.com"},
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "langchain-together>=0.3.0",
    "langchain-openai>=0.1.0",
    "langchain-deepseek>=0.1.0",
    "stanza>=1.10.0"
]

[project.urls]
"Homepage" = "https://github.com/SaveVic/llm_qus_analyzer"